requirements.txt
google-generativeai==0.3.1
streamlit==1.41.1
joblib==1.3.2
python-dotenv
pillow==10.2.0
pandas==2.2.1
numpy==1.26.4
plotly==5.9.0
matplotlib==3.7.2